# webapp/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Category API",
        "description": "API endpoints for managing categories",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/preorders": {
            "get": {
                "tags": [
                    "Preorders"
                ],
                "summary": "Get a list of preorders",
                "operationId": "23eaca1acfbd8ff06588e8cdeb578017",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Preorders"
                ],
                "summary": "Create a new preorder",
                "operationId": "e6d62f75f826d3e0e128ad7c5aa00dc3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "buyer_id": {
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "qty": {
                                        "type": "integer"
                                    },
                                    "deal_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "bill": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Preorder created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/preorders/{id}": {
            "get": {
                "tags": [
                    "Preorders"
                ],
                "summary": "Get a specific preorder",
                "operationId": "875a15a3421a7f486979bd43c891bf90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Preorder not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Preorders"
                ],
                "summary": "Update a specific preorder",
                "operationId": "c78fa9cbb05811c3e356049d7339762b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "buyer_id": {
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "qty": {
                                        "type": "integer"
                                    },
                                    "deal_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "bill": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preorder updated successfully"
                    },
                    "404": {
                        "description": "Preorder not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Preorders"
                ],
                "summary": "Delete a specific preorder",
                "operationId": "faf16dd09725d697fea29388317ea02c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Preorder deleted successfully"
                    },
                    "404": {
                        "description": "Preorder not found"
                    }
                }
            }
        },
...
    },
...
}